requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
tomli==2.0.1